refactor: add internal package for reusing test and other helpers
