+++ packages/internal/package.json
@@ -1,0 +1,15 @@
+{
+  "private": true,
+  "name": "@stacks/internal",
+  "scripts": {
+    "start": "tsc -b tsconfig.build.json --watch --verbose",
+    "typecheck": "tsc --noEmit",
+    "typecheck:watch": "npm run typecheck -- --watch"
+  },
+  "devDependencies": {
+    "@stacks/api": "^6.9.0",
+    "@stacks/blockchain-api-client": "^7.12.0",
+    "@stacks/network": "^6.16.0",
+    "jest-fetch-mock": "^3.0.3"
+  }
+}
