+++ packages/internal/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.json",
+  "include": ["src/**/*", "tests/**/*"],
+  "typedocOptions": {
+    "entryPoints": ["src/index.ts"]
+  }
+}
