+++ packages/stacking/package.json
@@ -32,6 +32,7 @@
   },
   "devDependencies": {
     "@stacks/blockchain-api-client": "^7.3.0",
+    "@stacks/internal": "file:../internal",
     "jest-fetch-mock": "^3.0.3",
     "process": "^0.11.10",
     "rimraf": "^3.0.2",
