+++ packages/stacking/tests/stacking-2.4.test.ts
@@ -1,23 +1,22 @@
+import { StacksNodeApi } from '@stacks/api';
 import { hexToBytes } from '@stacks/common';
 import { STACKS_MAINNET, STACKS_TESTNET } from '@stacks/network';
 import { ContractCallPayload, deserializeTransaction, getNonce } from '@stacks/transactions';
-
-import { decodeBtcAddress, StackingClient } from '../src';
-import { PoxOperationPeriod } from '../src/constants';
 import {
-  getFetchMockBroadcast,
   MOCK_EMPTY_ACCOUNT,
   MOCK_FULL_ACCOUNT,
   MOCK_POX_3_REGTEST,
-  setApiMocks,
   V2_POX_INTERFACE_POX_3,
   V2_POX_MAINNET_POX_2,
+  getFetchMockBroadcast,
+  setApiMocks,
   waitForBlock,
   waitForCycle,
   waitForTx,
-} from './apiMockingHelpers';
+} from '@stacks/internal';
+import { StackingClient, decodeBtcAddress } from '../src';
+import { PoxOperationPeriod } from '../src/constants';
 import { BTC_ADDRESS_CASES } from './utils.test';
-import { StacksNodeApi } from '@stacks/api';
 
 const API_URL = 'http://localhost:3999'; // default regtest url
