+++ tsconfig.json
@@ -9,6 +9,7 @@
       "@stacks/cli": ["packages/cli/src"],
       "@stacks/common": ["packages/common/src"],
       "@stacks/encryption": ["packages/encryption/src"],
+      "@stacks/internal": ["packages/internal/src"],
       "@stacks/network": ["packages/network/src"],
       "@stacks/profile": ["packages/profile/src"],
       "@stacks/stacking": ["packages/stacking/src"],
